+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use `clap` to parse arg
 - `HTTP Methods` of httpbin
 - `Request inspection` of httpbin
+- `Status codes` of httpbin
 - CI to check fmt, lint, build and test
 - CI to check `CHANGELOG.md`
 - CI to publish docker image